requirements.txt
streamlit
langchain-ollama
langchain-qdrant
qdrant-client>=1.7.0
langchain-community>=0.0.21
langchain>=0.1.0
pdfplumber>=0.10.0
ollama>=0.1.0
crewai>=0.28.0
flask==2.3.3
gradio==3.50.2
psycopg2-binary==2.9.7
openai>=1.68.2,<2.0.0
python-dotenv==1.0.0
requests==2.31.0
